civic_data_ingest/requirements.txt
# Data Collection & Web Scraping
tweepy==4.14.0
newspaper3k==0.2.8
beautifulsoup4==4.12.2
requests==2.31.0

# Database & Storage
pymongo==4.6.0
psycopg2==2.9.9
sqlalchemy==2.0.23

# NLP & ML
transformers==4.36.2
torch==2.1.1
keybert==0.8.1
vaderSentiment==3.3.2
scikit-learn==1.3.2
nltk==3.8.1

# Data Processing
pandas==2.1.4
numpy==1.25.2

# Visualization
plotly==5.17.0
folium==0.15.1
streamlit==1.29.0
matplotlib==3.8.2
seaborn==0.13.0

# Utilities
python-dotenv==1.0.0
geopy==2.4.1
pydantic==2.5.2
loguru==0.7.2